Allow for longer Post.guid
The maximum length of `Post.link` is currently 512 characters, but
`Post.guid` is only 255. Given post URLs are often used as guids,
this is a discrepancy.

This causes a fatal error when fetching a feed that includes a post with
a guid longer than 255 characters.

I've increased the `max_length` of `Post.guid` to 512 characters.
Hopefully that's long enough...

For #13